Moved custom JavaPlatform combo box code out from Gradle Compile customizer.
